+++ docs/releases.md
@@ -5,10 +5,13 @@
 ### New Features
 
 * Enhance `partition_netcdf` flexibility: now supports partitioning multiple input files and allows specifying a custom output directory for saved tiles ([#394](https://github.com/CWorthy-ocean/roms-tools/pull/394))
+* Option to plot bathymetry along fixed lat/lon section ([#399](https://github.com/CWorthy-ocean/roms-tools/pull/399))
 
 ### Breaking Changes
 
 ### Internal Changes
+
+* Require xgcm version >= 0.9.0 ([#425](https://github.com/CWorthy-ocean/roms-tools/pull/425))
 
 ### Documentation
 
@@ -48,6 +51,7 @@ plot(field, grid.ds)
 
 * Most classes now delegate their `.plot()` methods to the centralized plot function, reducing code duplication and simplifying maintenance. ([#375](https://github.com/CWorthy-ocean/roms-tools/pull/375))
 * Implement ruff rules ([#382](https://github.com/CWorthy-ocean/roms-tools/pull/382))
+* Implement streamable default datasources for GLORYS ([#386](https://github.com/CWorthy-ocean/roms-tools/pull/386))
 
 ### Documentation